pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rfid-servo-lock"
version = "0.1.0"
description = "A Raspberry Pi-based RFID-controlled servo motor lock system."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "Javid Ahmed", email = "javidahmed@icloud.com" }
]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "RPi.GPIO>=0.7.1",
    "mfrc522>=0.0.7",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-cov",
    "tomli",
    "validate-pyproject>=0.24.1",
]

[project.urls]
repository = "https://github.com/javidahmed64592/template-python"

[project.scripts]
rfid-lock = "rfid_servo_lock.main:run"
rfid-read = "rfid_servo_lock.rfid:read"
rfid-write = "rfid_servo_lock.rfid:write"

[tool.hatch.build]
include = [
  "README.md",
  "LICENSE",
  "rfid_servo_lock/**",
]

[tool.pytest.ini_options]
addopts = [
    "-vv",
    "--cov",
    "--cov-report",
    "term-missing",
]

[tool.ruff]
target-version = "py312"
line-length = 120
indent-width = 4

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

extend-exclude = [
    "**/__init__.py",
]

[tool.ruff.lint]
select = [
    "A",
    "ANN",
    "ASYNC",
    "B",
    "C4",
    "D",
    "E",
    "EM",
    "F",
    "FBT",
    "G",
    "I",
    "ISC",
    "LOG",
    "N",
    "NPY",
    "PD",
    "PERF",
    "PL",
    "PT",
    "RET",
    "RUF",
    "S",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "D203",
    "D213",
    "PLR0913",
    "S101"
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
warn_unused_configs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "RPi.*",
    "mfrc522.*",
]
ignore_missing_imports = true
